refactor to screen stories
