--- storybook/stories/Button.js
@@ -1,6 +1,0 @@
-import React from 'react'
-import Button from '../../src/components/common/Button'
-
-export default function () {
-	return <Button name="playlist-play" size={50} />
-}
