+++ storybook/stories/Explore.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Explore from '../../src/components/screens/Explore'
+
+export default function () {
+	return <Explore />
+}
