--- storybook/stories/Icon.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import Icon from '../../src/components/common/Icon'
-import { CenteredView } from '../CenteredView'
-
-export default function () {
-	return (
-		<CenteredView>
-			<Icon name="playlist-play" size={50} />
-		</CenteredView>
-	)
-}
