--- storybook/stories/MiniPlayer.js
@@ -1,8 +1,0 @@
-import React from 'react'
-import MiniPlayer from '../../src/components/common/MiniPlayer'
-
-export default function () {
-	const navigation = { navigate: () => 'navigate' }
-
-	return <MiniPlayer navigation={navigation} />
-}
