--- storybook/stories/OptionsBar.js
@@ -1,6 +1,0 @@
-import React from 'react'
-import OptionsBar from '../../src/components/screens/Player/OptionsBar'
-
-export default function () {
-	return <OptionsBar />
-}
