--- storybook/stories/PlayerControls.js
@@ -1,6 +1,0 @@
-import React from 'react'
-import PlayerControls from '../../src/components/screens/Player/PlayerControls'
-
-export default function () {
-	return <PlayerControls />
-}
