--- storybook/stories/PodcastsEmptyState.js
@@ -1,6 +1,0 @@
-import React from 'react'
-import PodcastsEmptyState from '../../src/components/screens/Podcasts/PodcastsEmptyState'
-
-export default function () {
-	return <PodcastsEmptyState />
-}
