+++ storybook/stories/Profile.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Profile from '../../src/components/screens/Profile'
+
+export default function () {
+	return <Profile />
+}
