--- storybook/stories/ProgressBar.js
@@ -1,6 +1,0 @@
-import React from 'react'
-import ProgressBar from '../../src/components/screens/Player/ProgressBar'
-
-export default function () {
-	return <ProgressBar />
-}
