+++ storybook/stories/Settings.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Settings from '../../src/components/screens/Settings'
+
+export default function () {
+	return <Settings />
+}
