+++ storybook/stories/SignIn.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import SignIn from '../../src/components/screens/Auth/SignIn'
+
+export default function () {
+	return <SignIn />
+}
