+++ storybook/stories/SignUp.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import SignUp from '../../src/components/screens/Auth/SignUp'
+
+export default function () {
+	return <SignUp />
+}
