+++ storybook/stories/Social.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import Social from '../../src/components/screens/Social'
+
+export default function () {
+	return <Social />
+}
