+++ storybook/stories/TopCharts.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import TopCharts from '../../src/components/screens/TopCharts'
+
+export default function () {
+	return <TopCharts />
+}
